requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-heroku==0.3.1
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0